manajemen user

+++ ROLE ADMIN/pages/crud_user/delete_pengurus.php
@@ -1,15 +1,19 @@
 <?php
-include '../../../Koneksi/koneksi.php';
+include "../../../Koneksi/koneksi.php";
 
-$id = $_GET['id'];
-
-$query = "DELETE FROM users WHERE id=$id";
-$result = mysqli_query($conn, $query);
-if (!$result) {
-    echo "SQL ERROR: " . mysqli_error($conn);
-    exit;
+if (isset($_GET['id'])) {
+    $id = mysqli_real_escape_string($conn, $_GET['id']);
+    
+    // Hapus data
+    $query = mysqli_query($conn, "DELETE FROM users WHERE id = '$id' AND role <> 'mahasiswa'");
+    
+    if ($query) {
+        header("Location: ../../index.php?page=manajemen_User&success_delete=1");
+    } else {
+        header("Location: ../../index.php?page=manajemen_User&error=delete_failed");
+    }
+} else {
+    header("Location: ../../index.php?page=manajemen_User");
 }
-
-header("Location: ../../index.php?page=manajemen_User");
-exit;
+exit();
 ?>
